ECOCSS-59: Đã fix lỗi dữ liệu và đồng bộ lịch sử commit

+++ cost-payment-service/src/main/java/com/example/costpayment/repository/CostRepository.java
@@ -5,8 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import java.time.LocalDateTime;
 import java.util.List;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
 
 @Repository
 public interface CostRepository extends JpaRepository<Cost, Integer> {
@@ -27,24 +28,24 @@ public interface CostRepository extends JpaRepository<Cost, Integer> {
     List<Cost> findByVehicleIdAndCreatedAtBetween(Integer vehicleId, LocalDateTime startDate, LocalDateTime endDate);
     
     // Tìm chi phí có số tiền lớn hơn hoặc bằng
-    List<Cost> findByAmountGreaterThanEqual(Double amount);
+    List<Cost> findByAmountGreaterThanEqual(BigDecimal amount);
     
     // Tìm chi phí có số tiền nhỏ hơn hoặc bằng
-    List<Cost> findByAmountLessThanEqual(Double amount);
+    List<Cost> findByAmountLessThanEqual(BigDecimal amount);
     
     // Tìm chi phí trong khoảng số tiền
-    List<Cost> findByAmountBetween(Double minAmount, Double maxAmount);
+    List<Cost> findByAmountBetween(BigDecimal minAmount, BigDecimal maxAmount);
     
     // Tìm chi phí theo Vehicle ID và khoảng số tiền
-    List<Cost> findByVehicleIdAndAmountBetween(Integer vehicleId, Double minAmount, Double maxAmount);
+    List<Cost> findByVehicleIdAndAmountBetween(Integer vehicleId, BigDecimal minAmount, BigDecimal maxAmount);
     
     // Query tùy chỉnh để tính tổng chi phí theo Vehicle ID
     @Query("SELECT SUM(c.amount) FROM Cost c WHERE c.vehicleId = :vehicleId")
-    Double getTotalAmountByVehicleId(@Param("vehicleId") Integer vehicleId);
+    BigDecimal getTotalAmountByVehicleId(@Param("vehicleId") Integer vehicleId);
     
     // Query tùy chỉnh để tính tổng chi phí theo loại
     @Query("SELECT SUM(c.amount) FROM Cost c WHERE c.costType = :costType")
-    Double getTotalAmountByCostType(@Param("costType") Cost.CostType costType);
+    BigDecimal getTotalAmountByCostType(@Param("costType") Cost.CostType costType);
     
     // Query tùy chỉnh để đếm số lượng chi phí theo Vehicle ID
     @Query("SELECT COUNT(c) FROM Cost c WHERE c.vehicleId = :vehicleId")
